requirements.txt
python-telegram-bot==20.7
supabase==2.3.4
google-generativeai==0.3.2
python-dotenv==1.0.0
vobject==0.9.7
pandas==2.1.4
openpyxl==3.1.2
